### 2019-08-17 #1658 @whedon
[PRE REVIEW]: LISC: A Python Package for Scientific Literature Collection and Analysis
**Submitting author:** @TomDonoghue (<a href="http://orcid.org/0000-0001-5911-0472">Thomas Donoghue</a>)
**Repository:** <a href="https://github.com/lisc-tools/lisc" target ="_blank">https://github.com/lisc-tools/lisc</a>
**Version:** v0.1.0
