### 2019-08-17 #1658 @danielskatz
👋 @TomDonoghue - thanks for this submission.  I'm going to assign myself as editor, and while I'll be slower than normal to respond over the next 2 weeks, I think we'll be able to make progress.

For your part, please look at this [list of potential reviewers](https://bit.ly/joss-reviewers) and identify individuals who might be able to review your submission (please start at the bottom of the list). Also, please suggest individuals who are not on this list by mentioning their GitHub handles here.  I also have some ideas for reviewers...
