### 2019-08-17 #1658 @danielskatz
👋 @pgroth - would you be willing to review this submission for JOSS?  (If you haven't done this before, which I think is the case, you might want to take a look at the [JOSS review guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and specifically [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html))
