### 2019-08-17 bb6e265
Change how grabs version, so gets same as it uses
