### 2019-08-19 #1658 @TomDonoghue
Hey @danielskatz - thanks for editing this submission!

I re-generated the PDF to fix some bullet lists that I hadn't quite formatted right.

In terms of reviewers, from the list of JOSS reviewers, the most relevant listings seem to me to be:
- Petar Maric (https://bitbucket.org/petar/)
- @\bagheria
- @\adammichaelwood
- @\linuxscout
- @\timClicks
- @\crew102

These aren't people I know personally, just the ones on the list for Python, who seemed to have the most related topic interests.
