### 2019-08-20 #1658 @danielskatz
Hi - I'm on vacation, so this will happen a little slower than normal (but still faster than a traditional journal), and thanks in advance for your patience
