### 2019-08-22 #1674 @danielskatz
👋 @timClicks, @linuxscout - your job as a reviewer is to examine the paper and software repo, and to check off items in your review list as possible.  When you have some issue where improvement is needed, either flag it here or with a new issue in the repository (mentioning this review issue).  If you have any questions or problems, please let me know. (FYI, I'm mostly on vacation this week and next and will be slower to respond than normal.)
