### 2019-08-23 #1674 @timClicks
Thanks @danielskatz. I'm also away until the middle of next week. I will aim to have my review completed by next Friday.
