### 2019-08-25 8c96ab1
Add cmap access to plot_matrix, & fix docs
