### 2019-08-25 864c057
Miscellaneous doc updates
