+++ lisc/analysis/counts.py
@@ -24,7 +24,7 @@ def compute_normalization(data, counts, dim='A'):
 
     Notes
     -----
-    - This computes a normalized data matrix as a percent of articles expressing co-occurrence.
+    This computes a normalized data matrix as a percent of articles expressing co-occurrence.
     """
 
     if dim == 'A':
@@ -66,8 +66,9 @@ def compute_association_index(data, counts_a, counts_b):
 
     Notes
     -----
-    - This computes a the Jaccard index, as AI_ij = |c_ij N d_ij| / |c_ij U d_ij|
-    - The denominator, |c_ij U d_ij|, is equivalent to |c_ij| + |d_ij| - |c_ij N d_ij|
+    This computes a the Jaccard index, as AI_ij = |c_ij N d_ij| / |c_ij U d_ij|
+
+    The denominator, |c_ij U d_ij|, is equivalent to |c_ij| + |d_ij| - |c_ij N d_ij|
     """
 
     n_a = len(counts_a)
