### 2019-08-25 06b93e2
Add file of requirements needed to build docs
