+++ requirements-docs.txt
@@ -1,0 +1,6 @@
+# Requirements for building the documentation
+sphinx
+sphinx_gallery
+sphinx_bootstrap_theme
+m2r
+numpydoc
