### 2019-08-25 f0ced5a
Switch format of API description to numpydoc
