+++ _sources/auto_tutorials/plot_04-WordsAnalysis.rst.txt
@@ -112,11 +112,11 @@ created from the aggregate data, across all articles for a given search term.
  .. code-block:: none
 
     frontal lobe :
-      Number of articles:            15
+      Number of articles:            14
       First publication:             2019
-      Most common author:            Lin CY
+      Most common author:            Silveira L
         number of publications:      1
-      Most common journal:           No shinkei geka. Neurological surgery
+      Most common journal:           World neurosurgery
         number of publications:      2
 
 
@@ -210,7 +210,7 @@ The Words object also has some methods for exploring the data.
 
  .. code-block:: none
 
-    <lisc.data.articles.Articles object at 0x114d15400>
+    <lisc.data.articles.Articles object at 0x133f4c8d0>
 
 
 
@@ -232,21 +232,21 @@ The Words object also has some methods for exploring the data.
 
  .. code-block:: none
 
-    Corpus callosotomy might have reduced epileptic seizure-induced repetitive shoulder joint dislocation in two patients with medically intractable epilepsy who were not focus resection candidates.
-    [White matter integrity in patients with mild cognitive impairment complicated with lacunar infarctions by diffusion tensor imaging].
-    Encapsulated Neuroprotection to the Rescue-or How to Safely Protect a Brain From Seizing.
-    Reality Is in the Posterior Hippocampus: Who Knew?
-    Adult Neurogenesis in the Development of Epilepsy.
-    What Is the Best Target for Ablation of Mesial Temporal Lobe Epilepsy?
-    Quantitative imaging biomarkers of damage to ritical memory regions associated with post-radiotherapy memory performance in brain tumor patients.
-    Neurons and networks in the entorhinal cortex: A reappraisal of the lateral and medial entorhinal subdivisions mediating parallel cortical pathways.
-    microRNA-132 is overexpressed in glia in temporal lobe epilepsy and reduces the expression of pro-epileptogenic factors in human cultured astrocytes.
-    Forward Processing of Object-Location Association from the Ventral Stream to Medial Temporal Lobe in Nonhuman Primates.
-    Between persistently active and activity-silent frameworks: novel vistas on the cellular basis of working memory.
-    Neuroanatomical abnormalities related to dexamethasone exposure in survivors of childhood acute lymphoblastic leukemia.
-    Tracking a changing paradigm and the modern face of epilepsy surgery: A comprehensive and critical review on the hunt for the optimal extent of resection in mesial temporal lobe epilepsy.
-    Intertemporal decision-making-related brain states predict adolescent drug abuse intervention responses.
-    Early ictal recruitment of midline thalamus in mesial temporal lobe epilepsy.
+    Voxel-based correlation of 18F-THK5351 accumulation and gray matter volume in the brain of cognitively normal older adults.
+    A 9.2-gram Fully-Flexible Wireless Ambulatory EEG Monitoring and Diagnostics Headband with Analog Motion Artifact Detection and Compensation.
+    Connectivity of the human insula: A cortico-cortical evoked potential (CCEP) study.
+    White Matter Topographic Anatomy applied to Temporal Lobe surgery.
+    Analysis of temporobasal vein with short subdural segment for anterior transpetrosal approach.
+    Anxiety and depression in patients with intracranial arachnoid cysts - a prospective study.
+    Cortical network underlying audiovisual semantic integration and modulation of attention: An fMRI and graph-based study.
+    [Isolated effective coherence analysis of epileptogenic networks in temporal lobe epilepsy using stereo-electroencephalography].
+    Olfactory function in focal epilepsies: Understanding mesial temporal lobe epilepsy beyond the hippocampus.
[…]
