+++ _sources/auto_tutorials/plot_00-Overview.rst.txt
@@ -402,7 +402,7 @@ can automatically load and save files to an organized output structure.
                 tutorial_counts.p
 
 
-**Total running time of the script:** ( 0 minutes  0.003 seconds)
+**Total running time of the script:** ( 0 minutes  0.002 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_00-Overview.py:
