### 2019-08-25 39b6463
Add note of pip download to README
