+++ _sources/auto_tutorials/plot_01-CountsCollection.rst.txt
@@ -256,7 +256,7 @@ For now, let's save out our collected counts data, using the LISC utility to sav
 
 
 
-**Total running time of the script:** ( 1 minutes  12.224 seconds)
+**Total running time of the script:** ( 1 minutes  6.769 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_01-CountsCollection.py:
