+++ _sources/auto_tutorials/plot_02-CountsAnalysis.rst.txt
@@ -287,7 +287,7 @@ collected co-occurrence data.
 
 
 
-**Total running time of the script:** ( 0 minutes  0.366 seconds)
+**Total running time of the script:** ( 0 minutes  0.376 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_02-CountsAnalysis.py:
