+++ _sources/auto_examples/plot_functions_counts.rst.txt
@@ -210,7 +210,7 @@ There are also plot functions available, same as demonstrated in the Counts tuto
 
 
 
-**Total running time of the script:** ( 0 minutes  13.418 seconds)
+**Total running time of the script:** ( 0 minutes  13.020 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_functions_counts.py:
