+++ _sources/auto_examples/plot_requester_urls.rst.txt
@@ -119,7 +119,7 @@ ensure requests respect API limits, as well as metadata collection, and URL logg
 
     Requester object is active:      True
     Number of requests sent:         3
-    Requester opened:                15:05:17 Sunday 25 August 2019
+    Requester opened:                15:08:30 Sunday 25 August 2019
     Requester closed:
 
 
@@ -140,7 +140,7 @@ ensure requests respect API limits, as well as metadata collection, and URL logg
 
  .. code-block:: none
 
-    {'is_active': True, 'n_requests': 3, 'wait_time': 0.5, 'start_time': '15:05:17 Sunday 25 August 2019', 'end_time': '', 'logging': None, 'log': None}
+    {'is_active': True, 'n_requests': 3, 'wait_time': 0.5, 'start_time': '15:08:30 Sunday 25 August 2019', 'end_time': '', 'logging': None, 'log': None}
 
 
 URLs Object
@@ -334,7 +334,7 @@ new API object by inheriting from and using the `URLs` object.
 
 
 
-**Total running time of the script:** ( 0 minutes  2.619 seconds)
+**Total running time of the script:** ( 0 minutes  2.578 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_requester_urls.py:
