+++ _sources/auto_examples/plot_functions_words.rst.txt
@@ -105,7 +105,7 @@ Here we will briefly explore collecting data directly using the function approac
 
  .. code-block:: none
 
-    [<lisc.data.articles.Articles object at 0x1280b0588>, <lisc.data.articles.Articles object at 0x1281daf60>]
+    [<lisc.data.articles.Articles object at 0x1259246d8>, <lisc.data.articles.Articles object at 0x1257c9208>]
 
 
 
@@ -153,7 +153,7 @@ check out the documentation for the `Articles` object, and what attributes it co
 
 
 
-**Total running time of the script:** ( 0 minutes  4.674 seconds)
+**Total running time of the script:** ( 0 minutes  4.702 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_functions_words.py:
