+++ _sources/auto_tutorials/plot_03-WordsCollection.rst.txt
@@ -105,7 +105,7 @@ and exclusion words, is used as was introduced and described in the first tutori
 
  .. code-block:: none
 
-    [<lisc.data.articles.Articles object at 0x13303cc18>, <lisc.data.articles.Articles object at 0x1335d5c50>]
+    [<lisc.data.articles.Articles object at 0x13063b6a0>, <lisc.data.articles.Articles object at 0x130bdaba8>]
 
 
 LISC Data Objects
@@ -253,7 +253,7 @@ We will analyze our words data in the next tutorial, so for now lets save out th
 
 
 
-**Total running time of the script:** ( 0 minutes  15.465 seconds)
+**Total running time of the script:** ( 0 minutes  15.630 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_03-WordsCollection.py:
