### 2019-08-25 f62dd04
Fix words plot to use Counter after data storage update
