+++ lisc/plts/words.py
@@ -36,7 +36,7 @@ def plot_years(years, year_range=None, ax=None):
 
     Parameters
     ----------
-    years : list of tuple of (int, int)
+    years : collections.Counter
         Data on the number of publications per year.
     year_range : list of [int, int], optional
         The range of years to plot on the x-axis.
@@ -47,8 +47,8 @@ def plot_years(years, year_range=None, ax=None):
     ax = check_ax(ax, (10, 5))
 
     # Extract x & y data to plot
-    x_dat = [xd[0] for xd in years]
-    y_dat = [yd[1] for yd in years]
+    x_dat = years.keys()
+    y_dat = years.values()
 
     # Add line and points to plot
     plt.plot(x_dat, y_dat)
@@ -57,9 +57,9 @@ def plot_years(years, year_range=None, ax=None):
     # Set plot limits
     if year_range:
         plt.xlim([year_range[0], year_range[1]])
-    plt.ylim([0, max(y_dat)+5])
+    plt.ylim([0, max(y_dat)+3])
 
     # Add title & labels
     plt.title('Publication History', fontsize=24, fontweight='bold')
-    plt.xlabel('Year', fontsize=18)
-    plt.ylabel('# Pubs', fontsize=18)
+    plt.xlabel('Year of Publication', fontsize=18)
+    plt.ylabel('Number of Articles', fontsize=18)
