### 2019-08-25 f19a389
Fix data type for plot inputs
