### 2019-08-31 997cbfc
Clean up spacing for when printing terms
