### 2019-09-01 #1674 @danielskatz
👋 @linuxscout - thanks for your work so far. Are there issues that are preventing you from checking the remaining boxes?  If so, please flag them here or with new issues in the repository (mentioning this review issue).
