### 2019-09-01 #1674 @linuxscout
I work on testing the software.
For the automated tests, there is none in the repository.
