### 2019-09-01 lisc-tools/lisc#22 @linuxscout
Automated tests
Are there automated tests or manual steps described so that the function of the software can be verified?
