### 2019-09-01 #1674 @linuxscout
No automated tests
https://github.com/lisc-tools/lisc/issues/22
