### 2019-09-01 #1674 @danielskatz
As we say in the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)

>Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
>
> * Good: An automated test suite hooked up to an external service such as Travis-CI or similar
> * OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g. a sample input file to assert behavior)
> * Bad (not acceptable): No way for you the reviewer to objectively assess whether the software works

So we do not require automated tests (though we encourage them), but we do require some tests.

In this case, if there are manual tests that you think are sufficient "to objectively check the expected functionality of the software (e.g. a sample input file to assert behavior)", that's enough for this criterion to be checked.
