### 2019-09-01 #1674 @linuxscout
I think that tutorials are good as tests.
