### 2019-09-01 #1674 @TomDonoghue
There are tests!

In the folder lisc/tests is a set of unit-tests. They are, I would say often smoke-tests, in that there is not always strong accuracy checking, but there are at least sanity check tests.

They are written with pytest. If you want to run them locally, then, with pytest installed, from inside the lisc code folder (so in `lisc/lisc`) run:
`pytest .`

I realize now it's not really documented anywhere how to run the tests. Is there a convention on how / where to note tests and how to run them?

Tests are also integrated with travis:
https://travis-ci.org/lisc-tools/lisc
