### 2019-09-04 d28f884
Add Dependency object to handle optional deps
