+++ lisc/tests/test_core_modutils.py
@@ -1,0 +1,11 @@
+"""Tests for the module utilities from lisc.core."""
+
+###################################################################################################
+###################################################################################################
+
+def test_dependency():
+
+    dep = Dependency('test_module')
+
+    with raises(ImporttError):
+        dep.method_call
