+++ lisc/tests/test_core_modutils.py
@@ -1,5 +1,9 @@
 """Tests for the module utilities from lisc.core."""
 
+from py.test import raises
+
+from lisc.core.modutils import *
+
 ###################################################################################################
 ###################################################################################################
 
@@ -7,5 +11,5 @@ def test_dependency():
 
     dep = Dependency('test_module')
 
-    with raises(ImporttError):
+    with raises(ImportError):
         dep.method_call
