### 2019-09-06 #1674 @danielskatz
> What I should do now?

Nothing - your part is done, and thanks!

We're now waiting for @timClicks to do his review, check off his items, bring up needed changes, etc.  He was away, but said he would have his review done around now.

I think we're also waiting for @TomDonoghue to add something about how to run the tests to the documentation.
