### 2019-09-08 #1674 @timClicks
@danielskatz My recommendation: accept, subject to the test amendments that have been mentioned already :)

I've worked through the tutorials, and think that this a very handy package.

In general, the API documentation is in an OK state, rather than a Good state (as per the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html)). That reflects the relative immaturity of the project. lisc is in a much better state than other packages that are released at a 0.1 state.

As a side comment, the tests are written using a very common test framework for Python applications (pytest). However, this isn't included as a dependency in the requirements.txt file so installing the test framework manually is required.

@TomDonoghue thanks for taking the time to submit this to JOSS.

---

@danielskatz @linuxscout @TomDonoghue Sincerely apologise for delaying this process.
