### 2019-09-08 #1674 @danielskatz
👋 @TomDonoghue - so it seem like the only remaining action for you is to add something about how to run the tests to the documentation.
