### 2019-09-09 19f8fa3
Mild tweaks to paper
