+++ paper/paper.md
@@ -23,24 +23,26 @@ bibliography: paper.bib
 
 # Summary
 
-The scientific literature is vast, and ever expanding. As a single example, the Pubmed database, a curated
-database of literature within the bio-medical sciences, holds more than 30 million published scientific articles,
-and continues to grow each year. Given the scale of the literature, work across informatics, information sciences, and
-bibliometrics has explored automated methods for the curation of and inference from the existing literature. This
-work is sometimes referred to as knowledge discovery, literature based discovery,
-or hypothesis generation [@stegmann_hypothesis_2003; @voytek_automated_2012; @spangler_automated_2014].
+The scientific literature is vast, and ever expanding. As a single example, the Pubmed database,
+a curated database of literature from the bio-medical sciences, holds more than 30 million
+published scientific articles, and continues to grow each year. Given the scale of the literature,
+work across informatics, information sciences, and bibliometrics has explored automated methods
+for the curation of and inference from the existing literature. This work is sometimes referred
+to as knowledge discovery, literature-based discovery, or hypothesis generation
+[@stegmann_hypothesis_2003; @voytek_automated_2012; @spangler_automated_2014].
 
-Here we introduce 'Literature Scanner', or 'LISC', an open-source Python module
-for performing automated meta-analyses of scientific articles, by collecting and analyzing
-data from the scientific literature. LISC seeks to provide an easily accessible interface that
-connects to the APIs from available databases, such as the Pubmed database, to collect literature data
-of interest. LISC also provides access to the OpenCitations database [@heibi_coci_2019], and provides
-an extendable approach that can be used to integrate additional APIs. LISC also includes support and
-utilities for analyzing the collected literature data.
+Here, we introduce 'Literature Scanner', or 'LISC', an open-source Python module for performing
+automated meta-analyses of scientific articles by collecting and analyzing data from the
+scientific literature. LISC seeks to provide an easily accessible interface that connects to
+external resources that make data available through application programming interfaces (APIs).
+For example, LISC connects to the Pubmed database, providing access to collect and analyze bio-medical
+literature, and to the OpenCitations database [@heibi_coci_2019] providing access to citation data.
+LISC is designed with an extendable approach that can be used to integrate additional APIs.
+LISC also includes support and utilities for analyzing the collected literature data.
 
 For data collection, LISC currently offers the following types of literature data collection:
 
-- Counts: tools to collect and analyze data on the co-occurence of specified search terms
+- Counts: tools to collect and analyze data on the co-occurence of specified search terms in the literature
 - Words: tools to collect and analyze text and meta-data for articles identified from a set of search terms
 - Citations: tools to collect and analyze citation and reference data
 
@@ -59,7 +61,7 @@ be expanded to included new databases, APIs and analyses as new dataset and tool
 
 The size and increasing scale of the scientific literature is prohibitively large for
 individual scientists to be able to keep up with. Common methods for literature summarization,
-including meta-analyses and systematic reviews require time-demanding manual work, and are often
+including meta-analyses and systematic reviews, require time-intensive manual work, and are often
 limited in scope and lag behind the literature. As a way to complement such approaches, multiple
 lines of investigation have shown how automated analyses of scientific literature can be applied
 to summarize and make inference from the existing literature [@stegmann_hypothesis_2003;
@@ -67,10 +69,10 @@ to summarize and make inference from the existing literature [@stegmann_hypothes
 
 Despite these established methods for analyzing the continuously growing literature, there is
 currently a relative lack of openly available tools to collect and analyze scientific literature.
-Although databases such as Pubmed have application programming interfaces (API), it still takes
[…]
