+++ paper/paper.bib
@@ -1,16 +1,20 @@
 
 @inproceedings{fox_confidence_2018,
-	title = {Confidence Levels in Scientific Writing: Automated Mining of Primary Literature and Press Releases},
-	booktitle = {Proceedings of the {Cognitive} {Science} {Society}},
+	title = {Confidence levels in scientific writing: automated mining of primary literature and press releases},
+	booktitle = {Proceedings of the 40th {Annual} {Conference} of the {Cognitive} {Science} {Society}},
 	author = {Fox, Will and Donoghue, Thomas},
 	year = {2018},
+	pages = {1680--1685},
+	publisher = {Cognitive Science Society}
 }
 
 @inproceedings{gao_automated_2017,
 	title = {Automated generation of cognitive ontology via web text-mining},
-	booktitle = {Proceedings of the {Cognitive} {Science} {Society}},
+	booktitle = {Proceedings of the 39th {Annual} {Conference} of the {Cognitive} {Science} {Society}},
 	author = {Gao, Richard and Donoghue, Thomas and Voytek, Bradley},
 	year = {2017},
+	pages = {2067--2072},
+	publisher = {Cognitive Science Society}
 }
 
 @article{heibi_coci_2019,
@@ -26,8 +30,8 @@
 	isbn = {978-1-4503-2956-9},
 	url = {http://dl.acm.org/citation.cfm?doid=2623330.2623667},
 	doi = {10.1145/2623330.2623667},
-	booktitle = {Proceedings of the 20th {ACM} {SIGKDD} international conference on {Knowledge} discovery and data mining},
-	publisher = {ACM Press},
+	booktitle = {Proceedings of the 20th {ACM} {SIGKDD} {International} {Conference} on {Knowledge} {Discovery} and {Data} {Mining}},
+	publisher = {ACM},
 	author = {Spangler, Scott and Myers, Jeffrey N. and Stanoi, Ioana and Kato, Linda and Lelescu, Ana and Labrie, Jacques J. and Parikh, Neha and Lisewski, Andreas Martin and Donehower, Lawrence and Chen, Ying and Lichtarge, Olivier and Wilkins, Angela D. and Bachman, Benjamin J. and Nagarajan, Meena and Dayaram, Tajhal and Haas, Peter and Regenbogen, Sam and Pickering, Curtis R. and Comer, Austin},
 	year = {2014},
 }
@@ -44,7 +48,7 @@
 @inproceedings{sybrandt_moliere_2017,
 	title = {MOLIERE: Automatic Biomedical Hypothesis Generation System},
 	author = {Sybrandt, Justin and Shtutman, Michael and Safro, Ilya}, 
-	booktitle = {Proceedings of the 23rd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+	booktitle = {Proceedings of the 23rd {ACM} {SIGKDD} {International} {Conference} on {Knowledge} {Discovery} and {Data} {Mining}},
 	series = {KDD '17},
 	year = {2017},
 	isbn = {978-1-4503-4887-4},
