### 2019-09-09 6e32c38
add notes on test coverage to README
