### 2019-09-09 adc0423
Shorten test notes
