+++ _sources/auto_examples/plot_requester_urls.rst.txt
@@ -96,7 +96,7 @@ ensure requests respect API limits, as well as metadata collection, and URL logg
  .. code-block:: none
 
     Collecting web page      https://www.google.com/         got status code         200
-    Collecting web page      https://www.yahoo.com/          got status code         200
+    Collecting web page      https://de.yahoo.com/?p=us      got status code         200
     Collecting web page      https://duckduckgo.com/         got status code         200
 
 
@@ -119,7 +119,7 @@ ensure requests respect API limits, as well as metadata collection, and URL logg
 
     Requester object is active:      True
     Number of requests sent:         3
-    Requester opened:                15:08:30 Sunday 25 August 2019
+    Requester opened:                23:29:40 Monday 09 September 2019
     Requester closed:
 
 
@@ -140,7 +140,7 @@ ensure requests respect API limits, as well as metadata collection, and URL logg
 
  .. code-block:: none
 
-    {'is_active': True, 'n_requests': 3, 'wait_time': 0.5, 'start_time': '15:08:30 Sunday 25 August 2019', 'end_time': '', 'logging': None, 'log': None}
+    {'is_active': True, 'n_requests': 3, 'wait_time': 0.5, 'start_time': '23:29:40 Monday 09 September 2019', 'end_time': '', 'logging': None, 'log': None}
 
 
 URLs Object
@@ -334,7 +334,7 @@ new API object by inheriting from and using the `URLs` object.
 
 
 
-**Total running time of the script:** ( 0 minutes  2.578 seconds)
+**Total running time of the script:** ( 0 minutes  2.792 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_requester_urls.py:
