+++ _sources/auto_tutorials/plot_06-MetaData.rst.txt
@@ -77,7 +77,7 @@ object, this collection information is attached and saved to the object, as the
 
  .. code-block:: none
 
-    2019-08-25_15:08:43
+    2019-09-09_23:29:55
 
 
 
