+++ _sources/auto_tutorials/plot_01-CountsCollection.rst.txt
@@ -120,10 +120,10 @@ The Counts object will now contain count data for the word co-occurrence data be
 
  .. code-block:: none
 
-    [[   0 1756  758  493]
-     [1756    0  709  597]
+    [[   0 1760  758  495]
+     [1760    0  709  598]
      [ 758  709    0  417]
-     [ 493  597  417    0]]
+     [ 495  598  417    0]]
 
 
 The Counts object also comes with some helper methods to explore the data.
@@ -148,10 +148,10 @@ The Counts object also comes with some helper methods to explore the data.
  .. code-block:: none
 
     The number of documents found for each search term is:
-      'frontal lobe'     -   13953
-      'temporal lobe'    -   27011
-      'parietal lobe'    -    4834
-      'occipital lobe'   -    3655
+      'frontal lobe'     -   13976
+      'temporal lobe'    -   27051
+      'parietal lobe'    -    4839
+      'occipital lobe'   -    3665
 
 
 
@@ -171,7 +171,7 @@ The Counts object also comes with some helper methods to explore the data.
 
  .. code-block:: none
 
-    The most studied term is  'temporal lobe'  with  27011  articles.
+    The most studied term is  'temporal lobe'  with  27051  articles.
 
 
 Counts: Two Terms Lists
@@ -256,7 +256,7 @@ For now, let's save out our collected counts data, using the LISC utility to sav
 
 
 
-**Total running time of the script:** ( 1 minutes  6.028 seconds)
+**Total running time of the script:** ( 1 minutes  36.176 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_01-CountsCollection.py:
