+++ _sources/auto_examples/plot_db_terms.rst.txt
@@ -245,7 +245,7 @@ term does not have an associated inclusion or exclusion term, leave that line bl
 
 
 
-**Total running time of the script:** ( 0 minutes  0.002 seconds)
+**Total running time of the script:** ( 0 minutes  0.004 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_db_terms.py:
