+++ _sources/auto_tutorials/plot_05-Citations.rst.txt
@@ -173,7 +173,7 @@ feel free to get involved with the project on `Github <https://github.com/lisc-t
 
 
 
-**Total running time of the script:** ( 0 minutes  4.453 seconds)
+**Total running time of the script:** ( 0 minutes  4.166 seconds)
 
 
 .. _sphx_glr_download_auto_tutorials_plot_05-Citations.py:
