### 2019-09-09 #1674 @TomDonoghue
**Test Coverage**

In terms of noting the test coverage, what I have done is add a section to the README, which gets built as the homepage of the documentation page, which mentions the tests, and how to run them, and the test dependency of pytest.

I'm actually not too sure what is considered convention / best practice for documenting tests & test coverage - so please let me know if this isn't the best answer! I haven't, for example, typically listed 'pytest' in the `requirements.txt` (it seems not necessarily standard to do so?). Would best practice be to list pytest in requirements, even if its not needed to use the package? Pytest is noted in setup.py. And/or, more generally, if anyone has any suggestions on additional or different ways to mention tests, please let me know!

**API Documentation**
I totally agree that the level of API documentation is middle ground, with it wanting for more examples with inputs / outputs, etc. That's definitely on the list of things to keep updating (https://github.com/lisc-tools/lisc/issues/20).

**Paper Notes**
There weren't any specific comments on the paper itself. I did a last sweep through, with some very minor edits for wording, layout, and minor ref fixes. I rebuilt the paper above. I hope it's okay to do some last tweaks - just wanted to make nothing was weird there before it gets locked in. Let me know if there are any comments on the wording, etc - happy to tweak / fix up!

**Thanks!**

A huge thank you to @linuxscout & @timClicks for doing the review, and @danielskatz for managing! Please let me know if there is anything else I should do!
