### 2019-09-09 #1674 @danielskatz
👋 @linuxscout & @timClicks - Please let me know your comments on the changes described above
