### 2019-09-10 #1674 @timClicks
@danielskatz LGTM. I've reviewed the new paper and the new README. Both look fine.
