### 2019-09-17 #1674 @danielskatz
👋 @linuxscout - Please let me know your comments on the changes described above
(last chance, then I'll assume this is ok with you based on your previous comments and your checking of all the items on your list)
