### 2019-09-26 408d60f
Fix Meta link & some typo fixes
