+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Summary
 
-The scientific literature is vast, and ever expanding. As a single example, the Pubmed database,
+The scientific literature is vast, and ever expanding. For example, the Pubmed database,
 a curated database of literature from the bio-medical sciences, holds more than 30 million
 published scientific articles, and is continuously growing. Given the scale of the literature,
 work across informatics, information sciences, and bibliometrics has explored automated methods
@@ -35,14 +35,14 @@ Here, we introduce 'Literature Scanner', or 'LISC', an open-source Python module
 automated meta-analyses of scientific articles by collecting and analyzing data from the
 scientific literature. LISC seeks to provide an easily accessible interface that connects to
 external resources that make data available through application programming interfaces (APIs).
-For example, LISC connects to the Pubmed database, providing access to collect and analyze bio-medical
+For example, LISC connects to the Pubmed database, providing access to collect and analyze biomedical
 literature, and to the OpenCitations database [@heibi_coci_2019] providing access to citation data.
 LISC is designed with an extendable approach that can be used to integrate additional APIs.
 LISC also includes support and utilities for analyzing the collected literature data.
 
 For data collection, LISC currently offers the following types of literature data collection:
 
-- Counts: tools to collect and analyze data on the co-occurence of specified search terms
+- Counts: tools to collect and analyze data on the co-occurrence of specified search terms
 - Words: tools to collect and analyze text and meta-data from scientific articles
 - Citations: tools to collect and analyze citation and reference data
 
@@ -50,12 +50,12 @@ To support use cases for collection and analyzing literature data, LISC includes
 
 - URL management and requesting for interacting with integrated APIs
 - custom data objects for managing collected data
-- a database structure, as well as save and load utilties for storing collected data
+- a database structure, as well as save and load utilities for storing collected data
 - functions and utilities to analyze collected data
-- data visualization functions for plotting collected data and analysis outputs
+- data visualization for plotting collected data and analysis outputs
 
 LISC is organized as an object-oriented tool, and aims to be a general utility that can
-be expanded to included new databases, APIs and analyses as new resources and tools are integrated.
+be expanded to included new databases, APIs, and analyses as new resources and tools are integrated.
 
 # Statement of Need
 
@@ -77,20 +77,19 @@ of interest and apply analyses of interest to it.
 LISC aims to serve as a complement to other relevant tools, for example Moliere,
 a more sophisticated and also more computationally complex tool for hypothesis
 generation [@sybrandt_moliere_2017], or Meta, a recently developed service for probing
-a pre-built knowledge network inferred from the literature
-(https://chanzuckerberg.com/science/programs-resources/meta/).
+a pre-built knowledge network inferred from the literature (https://www.meta.org).
 LISC, in contrast to these more complex systems, aims to offer a lightweight and customizable
 approach for finding and collecting literature of interest, and offers tools for efficiently
 performing analyses on this data. It aims to do so in particular by offering a connective
 interface between available APIs and natural language processing (NLP) analyses
 available through other tools. The goal is to allow for simple and rapid literature analyses.
-LISC may not be appopriate for more complex analyses and hypothesis generation projects
+LISC may not be appropriate for more complex analyses and hypothesis generation projects
 that would be best served by tools like Moliere.
 
 # Related Projects
 
 LISC is inspired by and based on the BRAIN-SCANR project, a project that collected literature
-data and analyzed co-occurences of terms in the neuroscientific literature [@voytek_automated_2012].
[…]
