### 2019-09-26 #1674 @TomDonoghue
@danielskatz - sorry for the delay, this time it's me away/travelling!

I have done a last check of the paper, and for the references, and other than a couple typos, the only thing I changed is the link to Meta (to be the tool link, not the page on Chan-Zuckerberg, which I think makes more sense). I also released a small version update of the tool, and tagged & released on GitHub / PYPI and made a release on Zenodo, to get a DOI.

In terms of what you asked for:
Current Version: 0.1.1
DOI: 10.5281/zenodo.3462035
Paper: I confirm the text & references are final

Please let me know if you need anything else from me!
