+++ _sources/auto_tutorials/plot_04-WordsAnalysis.rst.txt
@@ -114,10 +114,10 @@ created from the aggregate data, across all articles for a given search term.
     frontal lobe :
       Number of articles:            15
       First publication:             2019
-      Most common author:            Sahar F
+      Most common author:            Kim KM
         number of publications:      1
-      Most common journal:           World neurosurgery
-        number of publications:      2
+      Most common journal:           Journal of behavioral addictions
+        number of publications:      1
 
 
 Words Object
@@ -210,7 +210,7 @@ The Words object also has some methods for exploring the data.
 
  .. code-block:: none
 
-    <lisc.data.articles.Articles object at 0x133a50a20>
+    <lisc.data.articles.Articles object at 0x12308bb70>
 
 
 
@@ -232,21 +232,21 @@ The Words object also has some methods for exploring the data.
 
  .. code-block:: none
 
-    Fractional anisotropy of the optic radiations correlates with the visual field after epilepsy surgery.
-    APOE modifies the interaction of entorhinal cerebral blood flow and cortical thickness on memory function in cognitively normal older adults.
-    Phosphorylation of the HCN channel auxiliary subunit TRIP8b is altered in an animal model of temporal lobe epilepsy and modulates channel function.
-    Low-dose ladostigil for mild cognitive impairment: A phase 2 placebo-controlled clinical trial.
-    Case report: a giant arachnoid cyst masking Alzheimer's disease.
-    The infratranstentorial subtemporal approach (ITSTA): a valuable skull base approach to deep-seated non-skull base pathology.
-    Dynorphin-based "release on demand" gene therapy for drug-resistant temporal lobe epilepsy.
-    Antioxidant Capacity and Behavioral Relevance of a Polyphenolic Extract of Chrysanthellum americanum in a Rat Model of Irritable Bowel Syndrome.
-    Reactive Disruption of the Hippocampal Neurogenic Niche After Induction of Seizures by Injection of Kainic Acid in the Amygdala.
-    Epilepsy Mimicking Affective Disorder in a Patient with Amygdala Enlargement.
-    Environmental Enrichment and Brain Neuroplasticity in the Kainate Rat Model of Temporal Lobe Epilepsy.
-    High Frequency Oscillations in Epilepsy: Detection Methods and Considerations in Clinical Application.
-    A Rare Case of Insular Epilepsy: Not To Be Missed in Refractory Epilepsy Patients.
-    Biased Recognition of Surprised Facial Expressions Following Awake Craniotomy of a Right Temporal Lobe Tumor.
-    Assessing the effectiveness of perioperative s-ketamine on new-onset headache after resective epilepsy surgery (ESPAIN-trial): protocol for a randomised, double-blind, placebo-controlled trial.
+    New Approaches to Studying Silent Mesial Temporal Lobe Seizures in Alzheimer's Disease.
+    Genome-Wide Alteration of 5-Hydroxymethylcytosine in Hypoxic-Ischemic Neonatal Rat Model of Cerebral Palsy.
+    Dimethylethanolamine Decreases Epileptiform Activity in Acute Human Hippocampal Slices in vitro.
+    Trajectories of brain remodeling in temporal lobe epilepsy.
+    Dreamy State, Delusions, Audiovisual Hallucinations, and Metamorphopsia in a Lesional Lateral Temporal Lobe Epilepsy Followed by Ipsilateral Hippocampal Sclerosis.
+    Higher CSF tau levels are related to hippocampal hyperactivity and object mnemonic discrimination in older adults.
+    Examining the representational content of perirhinal cortex and posterior ventral visual pathway regions when maintenance of visual information is interrupted.
[…]
