+++ _sources/auto_examples/plot_functions_words.rst.txt
@@ -105,7 +105,7 @@ Here we will briefly explore collecting data directly using the function approac
 
  .. code-block:: none
 
-    [<lisc.data.articles.Articles object at 0x128636a58>, <lisc.data.articles.Articles object at 0x1285764a8>]
+    [<lisc.data.articles.Articles object at 0x122cbaba8>, <lisc.data.articles.Articles object at 0x122d12c88>]
 
 
 
@@ -139,13 +139,12 @@ Here we will briefly explore collecting data directly using the function approac
 
  .. code-block:: none
 
-    5 
+    4 
 
-    The tracking of lipopolysaccharide through the feto-maternal compartment and the involvement of maternal TLR4 in inflammation-induced fetal brain injury.
-    Speech-in-noise understanding in older age: The role of inhibitory cortical responses.
-    Intercomparison of MR-Informed PET Image Reconstruction Methods.
-    mage reconstruction for positron emission tomography based on patch-based regularization and dictionary learning.
-    N-acetylglucosaminyltransferase I promotes glioma cell proliferation and migration through increasing the stability of the glucose transporter GLUT1.
+    Amygdala Resting State Connectivity Differences between Bipolar II and Borderline Personality Disorders.
+    Protective Effect of Diphenhydramine against Traumatic Brain Injury in Rats via Modulation of Oxidative Stress and Inflammation.
+    Clinical Validation of Quality Improvements Using the Six Sigma Concept: A Case Study for Deep Brain Stimulation in Parkinson's Disease.
+    Plasma Levels, Temporal Trends and Clinical Associations between Biomarkers of Inflammation and Vascular Homeostasis after Pediatric Traumatic Brain Injury.
 
 
 To further explore the data collected and available, and what can be accessed,
@@ -153,7 +152,7 @@ check out the documentation for the `Articles` object, and what attributes it co
 
 
 
-**Total running time of the script:** ( 0 minutes  6.385 seconds)
+**Total running time of the script:** ( 0 minutes  6.367 seconds)
 
 
 .. _sphx_glr_download_auto_examples_plot_functions_words.py:
