### 2019-09-26 #1674 @danielskatz
@TomDonoghue - please see https://github.com/lisc-tools/lisc/pull/25
