### 2019-09-26 #1674 @TomDonoghue
@danielskatz - thanks for wording updates / fixes! I have merged the PR.
