### 2019-09-26 #1674 @danielskatz
@TomDonoghue - this seems to be taking a bit longer than normal to have the DOI work - It should work eventually...
