### 2019-09-26 #1674 @arfon
> 👋 @arfon - any insight on the DOI delay here?

Looks like it's resolving now.
